Move fileicon public tests to the right file
