+++ widget/fileicon_internal_test.go
@@ -12,7 +12,6 @@ import (
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/storage"
-	"fyne.io/fyne/v2/test"
 	"fyne.io/fyne/v2/theme"
 )
 
@@ -78,54 +77,6 @@ func TestFileIcon_NewURI_WithFolder(t *testing.T) {
 	item.SetURI(storage.NewFileURI(dir))
 	assert.Empty(t, item.extension)
 	assert.Equal(t, theme.FolderIcon(), item.resource)
-}
-
-func TestFileIcon_NewFileIcon_Rendered(t *testing.T) {
-	test.NewApp()
-	defer test.NewApp()
-
-	workingDir, err := os.Getwd()
-	if err != nil {
-		fyne.LogError("Could not get current working directory", err)
-		t.FailNow()
-	}
-
-	icon := NewFileIcon(nil)
-
-	w := test.NewWindow(icon)
-	w.Resize(fyne.NewSize(150, 150))
-
-	test.AssertImageMatches(t, "fileicon/fileicon_nil.png", w.Canvas().Capture())
-
-	text := filepath.Join(workingDir, "testdata/text")
-	icon2 := NewFileIcon(storage.NewFileURI(text))
-
-	w.SetContent(icon2)
-	w.Resize(fyne.NewSize(150, 150))
-	test.AssertImageMatches(t, "fileicon/fileicon_text.png", w.Canvas().Capture())
-
-	text += ".txt"
-	icon3 := NewFileIcon(storage.NewFileURI(text))
-
-	w.SetContent(icon3)
-	w.Resize(fyne.NewSize(150, 150))
-	test.AssertImageMatches(t, "fileicon/fileicon_text_txt.png", w.Canvas().Capture())
-
-	bin := filepath.Join(workingDir, "testdata/bin")
-	icon4 := NewFileIcon(storage.NewFileURI(bin))
-
-	w.SetContent(icon4)
-	w.Resize(fyne.NewSize(150, 150))
-	test.AssertImageMatches(t, "fileicon/fileicon_bin.png", w.Canvas().Capture())
-
-	dir := filepath.Join(workingDir, "testdata")
-	icon5 := NewFileIcon(storage.NewFileURI(dir))
-
-	w.SetContent(icon5)
-	w.Resize(fyne.NewSize(150, 150))
-	test.AssertImageMatches(t, "fileicon/fileicon_folder.png", w.Canvas().Capture())
-
-	w.Close()
 }
 
 func TestFileIcon_SetURI(t *testing.T) {
